+++ app/views/employees/show.html.erb
@@ -75,16 +75,8 @@
 
 <br>
 
-<h3>Total vacation days: <%= @all_vacations %></h3>
+<h3>Total vacation days: <%= @total_vacation_days %></h3>
 <br>
 <% @vacations.reverse.each do |vacation| %>
     <p><%= vacation.started_on %> - <%= vacation.finished_on %></p>
 <% end %>
-
-<%
-=begin%>
- <p>Vacation days: <%= @vacation_days %></p>
-<p>Used vacation days: <%= @used_vacation_days %></p>
-<p>Remaining vacation days: <%= @remaining_vacation_days %></p>
-<%
-=end%>
